+++ client/plugins/directusLogin.ts
@@ -1,0 +1,12 @@
+export default defineNuxtPlugin(async (nuxtApp) => {
+	const config = useRuntimeConfig()
+	const { login } = useDirectusAuth()
+	try {
+		await login({
+			email: config.public.directusEmail,
+			password: config.public.directusPassword,
+		})
+	} catch (error) {
+		console.error(error)
+	}
+})