+++ changelog.md
@@ -1,5 +1,11 @@
 # Nuxtus
 
- ## 0.0.1
+## 0.0.2
 
- Initial alpha release.
+- Move to using [nuxt-directus package](https://nuxt-directus.netlify.app/)
+- Add @nuxtus/cli
+- Add .DS_Store to gitignore
+
+## 0.0.1
+
+Initial alpha release.